# docs/swagger.json
{
  "components" : {
    "schemas" : {
      "Api-Error" : {
        "properties" : {
          "errorMessage" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "100 CONTINUE", "101 SWITCHING_PROTOCOLS", "102 PROCESSING", "103 CHECKPOINT", "200 OK", "201 CREATED", "202 ACCEPTED", "203 NON_AUTHORITATIVE_INFORMATION", "204 NO_CONTENT", "205 RESET_CONTENT", "206 PARTIAL_CONTENT", "207 MULTI_STATUS", "208 ALREADY_REPORTED", "226 IM_USED", "300 MULTIPLE_CHOICES", "301 MOVED_PERMANENTLY", "302 FOUND", "302 MOVED_TEMPORARILY", "303 SEE_OTHER", "304 NOT_MODIFIED", "305 USE_PROXY", "307 TEMPORARY_REDIRECT", "308 PERMANENT_REDIRECT", "400 BAD_REQUEST", "401 UNAUTHORIZED", "402 PAYMENT_REQUIRED", "403 FORBIDDEN", "404 NOT_FOUND", "405 METHOD_NOT_ALLOWED", "406 NOT_ACCEPTABLE", "407 PROXY_AUTHENTICATION_REQUIRED", "408 REQUEST_TIMEOUT", "409 CONFLICT", "410 GONE", "411 LENGTH_REQUIRED", "412 PRECONDITION_FAILED", "413 PAYLOAD_TOO_LARGE", "413 REQUEST_ENTITY_TOO_LARGE", "414 URI_TOO_LONG", "414 REQUEST_URI_TOO_LONG", "415 UNSUPPORTED_MEDIA_TYPE", "416 REQUESTED_RANGE_NOT_SATISFIABLE", "417 EXPECTATION_FAILED", "418 I_AM_A_TEAPOT", "419 INSUFFICIENT_SPACE_ON_RESOURCE", "420 METHOD_FAILURE", "421 DESTINATION_LOCKED", "422 UNPROCESSABLE_ENTITY", "423 LOCKED", "424 FAILED_DEPENDENCY", "425 TOO_EARLY", "426 UPGRADE_REQUIRED", "428 PRECONDITION_REQUIRED", "429 TOO_MANY_REQUESTS", "431 REQUEST_HEADER_FIELDS_TOO_LARGE", "451 UNAVAILABLE_FOR_LEGAL_REASONS", "500 INTERNAL_SERVER_ERROR", "501 NOT_IMPLEMENTED", "502 BAD_GATEWAY", "503 SERVICE_UNAVAILABLE", "504 GATEWAY_TIMEOUT", "505 HTTP_VERSION_NOT_SUPPORTED", "506 VARIANT_ALSO_NEGOTIATES", "507 INSUFFICIENT_STORAGE", "508 LOOP_DETECTED", "509 BANDWIDTH_LIMIT_EXCEEDED", "510 NOT_EXTENDED", "511 NETWORK_AUTHENTICATION_REQUIRED" ],
            "type" : "string"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Parking" : {
        "properties" : {
          "rates" : {
            "items" : {
              "$ref" : "#/components/schemas/Range"
            },
            "type" : "array"
          }
        },
        "required" : [ "rates" ],
        "type" : "object"
      },
      "Range" : {
        "properties" : {
          "days" : {
            "type" : "string"
          },
          "price" : {
            "format" : "int32",
            "type" : "integer"
          },
          "times" : {
            "pattern" : "^[\\d]{4}-[\\d]{4}$",
            "type" : "string"
          },
          "tz" : {
            "type" : "string"
          }
        },
        "required" : [ "days", "price", "times", "tz" ],
        "type" : "object"
      }
    }
  },
  "info" : {
    "title" : "Parking Prices - REST API",
    "version" : "v1.0"
  },
  "openapi" : "3.0.1",
  "paths" : {
    "/api/prices" : {
      "get" : {
        "operationId" : "getPrice",
        "parameters" : [ {
          "example" : "2011-12-03T10:15:30+01:00",
          "in" : "query",
          "name" : "start",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "example" : "2011-12-03T10:15:30+01:00",
          "in" : "query",
          "name" : "end",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Api-Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Api-Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Api-Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Get price for dates.",
        "tags" : [ "rates-controller" ]
      }
    },
    "/api/rates" : {
      "get" : {
        "operationId" : "getRatesInfo",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Parking"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Api-Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Api-Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Api-Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Retrieve rates.",
        "tags" : [ "rates-controller" ]
      },
      "put" : {
        "operationId" : "updateRatesInfo",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Parking"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Parking"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Api-Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Api-Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Api-Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Update rates.",
        "tags" : [ "rates-controller" ]
      }
    }
  },
  "servers" : [ {
    "url" : "https://localhost:5000"
  } ]
}